+++ src/pages/EventDetails/components/CountdownTimer.tsx
@@ -1,14 +1,11 @@
 import styled from '@emotion/styled';
 import { Box, Typography } from '@mui/material';
-import { differenceInMilliseconds, formatDistance, formatDistanceStrict } from 'date-fns';
+import { differenceInMilliseconds, formatDistanceStrict } from 'date-fns';
 import { nb } from 'date-fns/locale';
 import { useEffect, useState } from 'react';
-import { useQueryClient } from 'react-query';
 import { Link } from 'react-router-dom';
 
 import { Event, User } from 'types';
-
-import { EVENT_QUERY_KEYS } from 'hooks/Event';
 
 import Paper from 'components/layout/Paper';
 
@@ -49,9 +46,8 @@ type Order = {
   payment_link?: string;
 };
 
-const CountdownTimer = ({ user_id, event_id, payment_link, expire_date }: Order) => {
+const CountdownTimer = ({ payment_link, expire_date }: Order) => {
   const [timeLeft, setTimeLeft] = useState(convertTime(getTimeDifference(expire_date)));
-  const queryClient = useQueryClient();
 
   useEffect(() => {
     const interval = setInterval(() => {
